Update PartitionDevice's partedPartition when the PartedDisks get reset
We use 2 in memory copies of the partition table (2 PartedDisk instances),
one which we manipulate while auto / manual partitioning, and one
copy on which we reply the changes done by the auto / manual partitioning
step by step (committing each step to the disk) as we process our
actions.

For pre-existing partitions the PartitionDevice's partedPartition member
would point to the manipulated table's partitions, even though we had
already switched over and where actually executing actions. This causes
issues (segfaults!) with the flag clearing in ActionDestroyFormat.execute(),
as this was now calling unsetFlag on a partition which is no longer part of
a disk (as it was removed by clearParts).

This patch fixes this by making PartitionDevice's partedPartition property
switch over to the partitions in the new PartedDisk after we've
switched over to it.

p.s.

For some reason in my case, the segfault kept getting caught be some handler
and retriggered, so I had to scp in strace to find out it was segfaulting,
this showed up as the installer hanging after
"executing action: Destroy Format None on .... (partition)"

So this patch might fix some bug reports with these symptoms