Display bar view for the selected device only.
* iw/partition_gui.py (__init__): Add variable to track the displayed
storage element.
(getDisplayed, setDisplayed): New functions that manage the
varialbe holding the current displayed storage element.
(display): New function that handles the display of a disk.
(populate): Show only one disk in the bar view.
(treeSelectCB): When the user clicks on one of the disks or the
partitions, modify the bar view if needed.
(getScreen): A vertical Pane is not needed anymore.