Expose common fsset methods and properties in class Storage.
Add the following methods and properties to class Storage which map
through to FSSet methods, modify existing calls to use the method on
class Storage:

    turnOnSwap()
    mountFilesystems()
    umountFilesystems()
    parseFSTab()
    mkDevRoot()
    createSwapFile()
    fsFreeSpace()
    mtab
    mountpoints
    migratableDevices
    rootDevice

Callers no longer need to pass the Anaconda object to the FSSet methods
directly.  The method on Storage takes care of that.

The mtab() method on FSSet is exposed as a property on Storage.  The
same is true for fsFreeSpace().