Recheck if a partition should be ignored after getting its disk (#518971)
Recheck if a partition should be ignored after getting its disk (if its
disk was not scanned yet). Once we've the disk, we need to recheck if the
partition should be ignored again, as this catches cases when we scan the
partition before the disk and we have for example a disk which is part of
a BIOS raid set (and thus we should ignore the raw partitions).