Use type of device rather than name in booty target selection.
We were using startswith "md" together with checking the iswmd option to
determine wether an md array is a disk or a partition in the booty sense
of things. This is ofcourse wrong, instead we should check if it is an
isw mdraid array, or a native metadata mdraid array.

This patch fixed this, note the type of isw bios raid using mdraid is "disk".