Write mdadm.conf lines for mdraid container formats (imsm)
Ok, so the purpose of this patch is too write mdadm.conf lines for
mdraid container formats (imsm), but what it does is it cleans up
handling of mdraid container formats in general, with the
writing of mdadm.conf writing as a bonus effect really.

We were adding mdraid container members (raidsets inside the container)
to the device tree as DiskDevices. Which works, but is not completely
correct. This patch introduce a PartitionableMDRaidArrayDevice and
uses that for mdraid container members instead. This means we now also
correctly tear them down / set them up when asked. In the future (mostly
needs UI work to export the functionality) PartitionableMDRaidArrayDevice
can be used to create partitionable normal (native metadata) mdraid sets too.