Collect all modules from modules.{ccwmap|networking} on s390x (#522519)
Collect all modules listed in modules.ccwmap and modules.networking for
inclusion in the initrd.img for s390x.  Removed some module names from
mk-images.s390 that don't appear to exist anywhere in the kernel RPM
anymore.  Use $S390EXTRAMODS to list modules for inclusion that are not
in a mapfile.

Added expandModuleDeps() to make sure dependent modules are collected.