Don't fail to commit partitions due to active lvm/md.
If we get a failure to commit a partitioning change to disk, the first
thing to do is try shutting down all devices: In the case of lvm-on-md
we implicitly activate the md device in order to wipe an lv that we're
about to destroy. Subsequent attempts to commit to a disk containing any
of the md device's members will fail because the md is still active.