Add 'install' user to start anaconda on s390x.
This is a usability improvment.  Once you've booted up the kernel and
initrd on s390x, you are told to ssh in to the host to start the
installation.  In the past, we had rigged the root account to launch
loader on the first login and a shell on subsequent logins.  In this
patch, I've added an 'install' user that has loader for the shell.  The
root user still has bash for the shell.  Also change the recommendation
to run 'ssh -X' rather than 'ssh -x'.