Move S390MODS to inside makeBootImages(), remove libiscsi_tcp.
Since s390x is building a list of arch-specific modules for the
initrd.img based on what exists in /lib/modules/VERSION/kernel/s390, we
need to build the list inside makeBootImages() so that all required
variables are set up and the kernel is unpacked.

Remove libiscsi_tcp from the list since the iscsi modules are included
when INITRDMODS is expanded in mk-images.