Control the sensitivity of the "delete" and "create" buttons
When users click on stuff that is not editable or deletable we should
not let them click on the edit and/or the delete buttons.  It is still
possible to click edit or delete for a partition that is part of a
Volume Group or a RAID device.  In this case the user will be presented
with a message that describes the reason for not editing or deleting.
This behavior is on purpose so the user is not confused when the
"delete" or the "edit" button are not active on something that could be
thought as editable or deletable.  Also, I want to avoid going down to
the storage class every time the user clicks on the GUI.
Now we are sure that we receive a device in these functions and
therefore don't need checks in deleteCB and editCB.

* iw/partition_gui.py (treeSelectCB):  set_sensitive(True) when user
clicks on something that is editable or deletable.
(getScreen): make the edit button and the delete button part of the
class variables so we can access them in treeSelectCB.
* iw/partition_gui.py (deleteCB, editCB): Stop checking to see if what
is selected is a device.