Update po/anaconda.pot during a 'bumpver' run.
Since we have to store anaconda.pot in the source tree for the
translators, make sure we update it when bumpver is run so it stays
current.

When you run 'make bumpver', you will now be checking in at least two
files but up to three:

    anaconda.spec.in    [required]
    configure.ac        [required]
    po/anaconda.pot     [optional, only if it needed updating]

Make sure you add all of these files for 'New version' commits.