Add the create LV option.
* iw/datacombo.py (set_active_text):  Make sure we return False when we
don't find 't'.
* iw/partition_gui.py (createCB): Activate the 'create lv' radio button
only when there is free space in a Volume Group.  Create a Combo box
with a list of the Volume Group(s) that have free space.  Pre-select the
Volume group that the user has selected in the tree view.
(editCB): Call the editLVMLogicalVolume with its new parameters.
(editLVMLogicalVolume): Change the parameters for this function.
(device, isNew = False) -> (lv = None, vg = None).  In this way if lv is
none and vg is a Volume Group we are creating a new LV in the vg Volume
Group.  If lv is not None, we are editing lv.
* ui/create-storage.glade: Put the 'create lv' radio button in a
Horizontal box so we can fit the Volume Group Combo Box.