Correct makeupdates script to work with deleted files.
This patch avoids a traceback when trying to copy a deleted file.

* scripts/makeupdates (doGitDiff): used 'git diff --name-status tag' to get
the list of changed files.
(pruneFile, isysChanged):  Adjust for file name parsing.  We should stop
looking for the '|' character and we should analyze the first character
which contains the status of the file with respect to tag.  For now we
only skip the deleted files.