Create getCurrentDeviceParent function.
One of the ways of knowing where we are in the tree view is to ask for
the parent of the current row.  If there is no parent, then we know that
we are in a root row.  Additionally, when we click on free space, we
need the parent to know what to put in the bar view.
* iw/partition_gui.py (getCurrentDevice): Access the PyObject directly
from the model (model[iter]["PyObject"]).  If there is not PyObject then
it will contain None.
(getCurrentDeviceParent): Use model.iter_parent(iter) to get the parent
from iter.  Will return None if no parent is found.