Stop dragging mkinitrd into the install
We were forcebly installing mkinitrd, but nothing uses it anymore now that
we've switched to dracut. There is no need to force installing dracut as the
kernel requires it. And if someone decides to not install the kernel, well
though luck buddy.

p.s.

This patch is intended for both devel and F-12