Pass only cCB and dcCB to the StripeGraph classes.
The stripe graph classes do not need the tree object and should treat
the editCB as a double click call back (dcCB).  This is all done to be
more consistent with the logic behind the graph classes.

* iw/partition_gui.py (DiskStripeGraph, LVMStripeGraph)
(MDRaidArrayStripeGraph): Remove the tree and editCB as arguments.  Add
one click call back (cCB) and the double click call back (dcCB) to the
__init__ method.
(treeSelectCB): Pass the self.editCB as dcCB.  Pass the
self.selectRowFromObj as cCB.