Implement the double click for free space on the bar view
When double clicking on a free space from a 'normal' device, anaconda
should display a dialog to create a new partition.  When double clicking
on a free space from a volume group, anaconda should display a dialog to
create a new logical volume.

* iw/partition_gui.py (getSelectedSlice): New function that returns the
slice that is selected in the stripe.
(_createStripe): Always pass the self.cCB and self.dcCB to the Slice
class.
(barviewActivateCB): New function to handle the double click on the bar
slices.
(treeSelectCB): Make sure to pass the new barviewActivateCB function
instead of the editCB function to the stripeGraph objects.