Take 70-persistent-net.rules generated at installation (#526322)
Discovered on s390x, but really affects all platforms.  I've done test
installs of RHEL-6 on s390x and of rawhide on x86_64 with this patch in
place and 70-persistent-net.rules looks correct on the target system
after installation.

The problem on s390x was we needed to use the KERNELS setting for the
line in the rules file, but on other platforms we don't need that.
Since /lib/udev/write_net_rules is kicked off during installation, copy
/etc/udev/rules.d/70-persistent-net.rules from the installation
environment to the target system.  If we don't have that file in the
installation environment, just do what we have been doing (which should
take care of the instances where people are doing kickstart installs
with multiple network lines but onboot is set to no).