Make storage.devices.deviceNameToDiskByPath() more robust.
Have deviceNameToDiskByPath() return the full device path rather than
just the basename path.  Have it check to see if the entry in
/dev/disk/by-path is actually a symlink, then readlink that.  Otherwise
just take the basename of the entry.  Also fix a problem where passing
it a full device path would not match (e.g., '/dev/dasdb1').  Take the
basename of deviceName and match that against the readlink() value.