Reset parted flags in createFormat not destroyFormat
We were resetting the parted flags for re-use of existing partitions in
destroyFormat, but this assumes that the existing partition actually has
a type in the partition table matching its current formatting, this has 2
problems:
1: If the partition has a type which needs to have flags reset to become
   the correct type for the new format (for example swap -> ext2), but it
   is not formatted, the flags wont get reset
2: If the partition has a type which does not match the formatting (for
   example type linux raid, formatting ext2), the flags won't get reset

This patch moves the resetting of the flags to the creation of the new
format, and resets all flags except for the one for the to be created
format (and boot and lba which are special).

For the F-12 branch this also has the benefit that the resetting of the flags
happens at all, as there the resetting was happening on a deep copy of the
device (and thus the label) and never actually got written to the disk.