Set partedPartition system to the correct FS when creating an FS
The playing around with partedPartition flags only allows us to determine
the partition table entry for certain usuages which are deemed special
by parted. For normal FS usuage, parted will default to the type for the FS
it has detected (for pre-existing partitions) or to Linux (83) for new
partitions.

This means that for example reformatting a vfat partition as ext3, or a new
partition as vfat will lead to incorrect partition type entries in the
partition table. This patch fixes this.

This patch is intended for both master and F-12.