Use addUdevPartitionDevice() for adding dmraid / multipath partitions (#527785)
With a kickstart install, with "clearpart --label", we will end up
scanning pre-existing partitions after clearing the label. This leads
to the backtrace as seen in #527785 and #526517.

addUdevPartitionDevice() has special handling for this case, instead of
copy and pasting this, this patch modifies addUdevPartitionDevice() so
that it can be used for dmraid and multipath partitions too.