Default to mbr bootloader target for mdraid 1 boot device too (#526822).
It used to be so, because for boot device target (/dev/mdX) we installed into
MBRs anyway. With Hans's patch 45a7048e5f56316e052e4699b5ec70aa291ddd5e we do
what we promise (install to boot device) in this case, but it is a change and
it doesn't seem to be expected as default by users.