Handle Installation Repo (base repo) as any other in repo edit UI.
With present UI, edit repos, not base repos (methodstr).  Works with
assumption that base repo (repo/method) url points to a single repo, which is
true for Fedora (as opposed to RHEL5 where we various repos in directories).
dgregor told me that the target for RHEL6 is to have just one repo and handle
product differentiation by comps. If we had to work with multirepo base
urls, the UI for repo editing would need to be redesigned and rewritten because
now we are mixing base urls with repo urls there.