Add dracutSetupString() method to ZFCPDiskDevice (#526354).
Collect CCW bus ID, WWPN, and FCP LUN values for zFCP devices when
building the device tree.  Store these in the ZFCPDiskDevice object and
use them to generate the rd_ZFCP= string for dracut.

Expand storage/udev.py with functions to determine if a device is zFCP
and to get arbitrary attribute values.