Make bootLoaderInfo new-style class, so that its properties work correctly.
If the bootLoaderInfo is old-style class, setting of its property drivelist
doesn't work, instead the property is overwritten with attribute and therefore
for example any following calls of updateDriveList have no effect.

AFAIK it doesn't cause problems anywhere at the moment, but I think it should
be fixed it in master.