Improve message given to user for fsck failures (#527626).
If fsck fails with a return code indicating errors we can't
automatically fix, give the user a more descriptive error dialog before
exiting.  There's not much we can do if the filesystem is beyond
automatic repair, so tell the user to repair the filesystem under rescue
mode or something similar.