Find and format any unformatted DASD devices (#528386).
Before the device tree is built in the storage code, run through the
DASD devices on the system and format any with a status of
'unformatted'.  We have to run dasdfmt on these devices before we can
partition them.

Under RHEL-5, this process was serial and the user had to click Yes for
each unformatted DASD found, which could take a really long time if
you had thousands of DASDs.

The idea now is that if the DASD is seen by anaconda, we want to use it
for installation.  The stage 1 device initialization routines as well as
the CMS conf file provided at boot time allow the user to restrict the
range of devices we see during installation.  If any of the devices we
see are unformatted, run dasdfmt before building the devicetree.

This process is implemented in a DASD class that works similar to the
iscsi and zfcp classes.  That is, devices may need to be started or
brought in to a working state before the anaconda storage code can use
them.