Fix DASD and zFCP device discovery (#530287).
Syntax error with kwargs in DASDDevice.  Use kwargs.pop() in both
DASDDevice and ZFCPDiskDevice so DASD or zFCP specific kwargs are not
passed to DiskDevice.__init__().

Remove the incorrect name variable setting in ZFCPDiskDevice.

Change 'devices' to 'device' in udev_device_get_zfcp_attribute() so that
zFCP attributes can be read.