Improve detailedMessageWindow() in text.py.
Previously, detailedMessageWindow() would just give you a messageWindow
and throw out the longText.

This patch tries to implement something that will let the longText
contain somewhat formatted text and allow users to scroll.  That's done
with the ListboxChoiceWindow.  Selecting any item in the list box is the
same as selecting button ID 0 on the dialog.