Clear pot and po updates after a 'make release' or 'make archive'.
These are updated by msgfmt, but leave the po files in the repository as
is since they are updated by an external project.  We'll keep the pot
file up to date, but that happens at 'make bumpver'.  Since msgfmt
generates new po files for the release tarball, we need to reset all of
these files after a 'make release' or 'make archive'.