Add 'zerombr' to list of early kickstart commands we look for.
zerombr is used during storage.DASD.startup() if the user is performing
a non-interactive kickstart install.  Since the full kickstart file is
not processed until after the storage object has initialized, we need to
include this flag in the early kickstart processing because it's used
during the storage object initialization.