Modify autopart requests to include a separate /home (#150670).
This modifies the default partitioning as follows:

(1) For VGs <= 50 GB, we will continue to make swap and / as normal.
(2) For VGs >= 50 GB, / will cap at 50 GB and /home will consume the rest.

50 GB is fairly arbitrary, and was based on the fact that an Everything
install of Fedora right now is ~ 40 GB, plus some room for expansion.  Very
few users are likely to do an Everything install so this should provide plenty
of space for upgrades and future growth.  Additionally, this is only a default
partitioning suggestion and can always be overridden by the user.