Fix "resize failed: 1" errors for ext2/ext3/ext4 (#517491).
The following install test case has been failing:
https://fedoraproject.org/wiki/QA:Testcase_Anaconda_autopart_%28shrink%29_install

The problem was with the minSize property in Ext2FS.  We use resize2fs
to get the minSize for ext2/3/4 filesystems, which is good because
resize2fs accounts for additional things an extX volume may need.  The
problem is the value it reports is in blocks.  We have to convert those
blocks to bytes, then to megabytes, then round up to account for any
fractional megabytes.  Use dumpe2fs to get the block size and use
resize2fs as we have been, but modify the calculation of size.

Also the _setTargetSize() method in FS needed a change.  minSize can be
less than or equal to newsize, not just less than.