Move identifyMultipaths from DeviceTree to devicelibs.
The return type has also changed.  It now returns a tuple of a list of regular
devices, a list of multipath sets, and a list of partition devices.  This is
necessary so the filtering UI can use the same knowledge that was in the
DeviceTree object to know which devices are part of a multipath set and which
are just standalone disks.