Correct modopts initialization in loader (take 2) (#531932).
The edc665e6fa2ba71e89eb83412738622e916c3a05 commit prevented SIGSEGV
but changed behavior of modopts to where we'd never get any values in
modopts.

What needed to happen is the else clause in addOption() needed to
initalize the .options array and add in the option value.  That wasn't
happening, so strcat() calls in writeModulesConf() were causing SIGSEGV.

This patch also adds some realloc() and malloc() checks.  Tested with
"radeon.nomodeset=1" and without that parameter on x86_64.