Initialize DASD totalCylinders before progress bar callback (#532420).
On s390, if you had a single DASD that needed dasdfmt run, you would get
a ZeroDivisionError traceback because the self._totalCylinders value was
zero.  The problem was caused by the DASD.totalCylinders property and
how it initialized.  It would initialize itself on the first call, but
since we'd already fired off a dasdfmt process for the device, the one
running to capture the cylinder count couldn't get access to the device.

Users with more than one DASD needing a dasdfmt would not see the
traceback, but would have an incorrect total cylinder count, so 100%
would be reached when n-1 devices had been formatted.

Lastly, in certain instances, the status for the DASD could be "basic"
rather than "unformatted", so treat those the same and perform a
dasdfmt.