Convert string.find calls into something modern
string.find('foo') != -1 is a bit klunky.  We now have the ability to
just do 'foo' in string and get back a True or False from that.  I'm
sure there are more files in anaconda that could use this treatment but
this is the file I was looking at anyway.