Include the command line to put anaconda into debugger mode in history.
Things being in the history means I don't have to type them.  This
particular thing is fairly specific to get right, so it's a really good
candidate to go in there.  So this patch adds it.

This also changes the "killall" that's already there to use the pid
file, so we know which process it'll pick.