Add a dialog to configure advanced storage devices.
This brings back the old behavior of having a dialog that can prompt for
unusual storage devices that require manual intervention, like FCOE and iSCSI.
After the dialog is run, we have to put and new devices into the UI.  However,
udev isn't going to provide just a list of newly appeared devices so we have to
maintain a list of what was around previously, and compare the current device
list to that.  This promises to be slow but there's not a better option.