Move libcurl initialization to urlinstTransfer() (#537870).
Recent trees have been giving errors when we get to the stage2 download
point, such as this one:

    Unable to retrieve http://download.fedoraproject.org/pub/fedora/linux/development/x86_64/os//images/install.img.

libcurl needs to be initialized inside urlinstTransfer() so it has the
latest network state for the system.  The above error is caused by old
DNS settings.  libcurl won't automatically do a res_init(), so we need
to set up a new curl instance and use that.

Since this patch moves curl usage to be exclusively within
urlinstTransfer(), move 'curl' to that function as well and remove it
from loaderData.

Add curl cleanup calls at the end of urlinstTransfer().