Sleep if the kickstart file read fails (#537361)
If a read from a block device (like USB key) or a cdrom (USB cdrom) fails, sleep
a bit to give the device some time to initialize. Return status of the mount
command is checked so waiting occurs only if there is a reasonable chance that
the device might eventually come online. This required extending our existing
mounting methods to report the correct mount error code.

I tested that
* the installer still runs
* reading the kickstart from a usb works
* starting the installer with no USB key inserted but a command line argument pointing to a USB drive will trigger the waiting cycle and that when the key is inserted while waiting the kickstart is eventually found on the key.