Use installclass to make the bootloader timeout 5 seconds on RHEL.
In RHEL, we want the old bootloader timeout behavior, so people see
output on the teletype and have time to frob some switches on the
front panel.  But for those of us in modern environments, we don't want
to do that, so this patch uses installclass to choose which one you get.