If /boot is too small to preupgrade, don't allow going back (#499321).
First, there's nothing you can do.  We don't allow changing the package set
or resizing partitions on upgrade, so you can't hope to recover.  Second,
trying to go back is bringing up a tricky bug.

On preupgrade, all the steps before this one are in skipSteps.  So dispatcher
tries to find the step to go back to, discovers there isn't one, and decides
to move "back" to the first step that wasn't skipped.  To no one's surprise,
that's the same step we were on.  However it's not an interactive step, so
setScreen tracebacks.