stage1 (init): Make /tmp tmpfs large enough to hold install.img (#540146)
Be default mount makes tmpfs mounts 1/2 RAM size, on 256MB machines, where
part of the RAM is taken up by a shared framebuffer, this results in a
/tmp which is to small to hold install.img, resulting in install failure
when trying to do for example a network installation.

This patch fixes this by specifying a size for /tmp, it uses 250MB which
should be plenty for now and the forseeable future.