Remove /etc/localtime before trying to copy into it (#533240).
If the file is already linked with the source the raised inspection kills
anaconda. This happens in some unclear livecd situations but in general nothing
prevents the user or any software he runs on the livecd from arranging this
manually.

Tested when the file doesn't exist beforehand (the common scenario) and when the
file links to its later origin already (the bug scenario).