Change btrfs command line option
In F11 and F12 the command line option to enable btrfs during the installer was
"icantbelieveitsnotbtr".  Unfortunately I can't spell believe the first time
every time, and btrfs has come quite a ways since F11 and is stable enough to
start being tested on a larger scale.  So change the command line option to
"btrfs" in keeping with tradition and to make it a bit simpler for those who
want to take it for a test spin in F13.  Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>