Make the fcoe, iscsi and zfcp classes singletons
The fcoe, iscsi and zfcp classes are designed to be instantiated only once,
currently this is not enforced. This is causing trouble as their single
instantiation currently is done in storage.__init__, which means they cannot
be used before instData is instantiated, as that instantiates storage.

However we need them to connect to fcoe, iscsi and/or zfcp drives while
parsing kickstart files (so the drivers can be referenced by UUID /
/dev/disk/by-path / whatever), and instData is not yet instantiated at that
time.

So this patch uses a very simple Singleton design pattern, so that we can
instantiate them where ever we need them, and all references returned will
will point to one shared global instance.