kickstart: Move onlining of fcoe/iscsi/zfcp devices to parse phase
Bring fcoe/iscsi/zfcp devices online when parsing the kickstart file
(instead of delaying this to the execute phase) so that the disks
can be identified by /dev/disk/by-path/foo, etc. in clearpart, ignoredisks,
etc. commands.

For the iscsi command this also improves error handling by passing
username and password to addTarget as is instead of only passing them if
both are present, using addTarget's error handling for a username without
a password (instead of simply ignoring the username or the password), and
also by catching IOErrors which will be thrown when logging in to the target
fails.