Provide CMS script for IPL under z/VM (#475343).
The installation docs tell users to write a REXX boot script to get the
RHEL installation going on s390x.  IBM wants that sample REXX script
provided in the images/ subdirectory.  This patch is a backport of two
patches from RHEL-6 (the one that introduces redhat.exec and the one
that modifies the script to reference generic.prm).