Avoid duplicate kernel cmdline options and cleanup booty dracut code
It is possible for 2 disks to have the same disk dracut cmdline options,
or network cmdline options (if the disks are network devices) for example
2 iscsi disks on the same host. In this case we should only include the
specific option once.

This patch puts the collection of storage related setupStrings in a function,
and makes get() loop over all kernel cmdline providing bits, so that the
code for proper whitespace seperation of various args is not duplicated.