Bump the required version numbers on a couple of components.
Whenever making an update in anaconda that requires some new feature in a
library, please consider updating anaconda.spec.in to state the correct
version number.  This didn't used to be so important, but in a world where
we are doing our own composes for multiple trees, it's really helpful.