Ask about LVM inconsistencies only in storageinit step.
The patch does essentially the same thing as commit
31d5b8f4fec513e4f8e246e47e9fee75c701d07b for disk initialization question. The
difference is that when user chooses LVM reinitialization, it is done
immediately (not as planned action as with disk initialization) so the user
wouldn't be asked again anyway without the patch (he would be asked if he chose
to ignore).

I also made one function more readable in the patch.