Don't wait on the filtertype screen on kickstart installs.
Since ignoredisk is an optional command and leaving it out of a kickstart
file means you don't want to ignore anything, there's no reason to stop
on that screen and wait for input if ignoredisk was not specified.  However,
if this is an interactive kickstart install we do still need to stop and
ask.