mdraid: various changes to options for new mdraid array creation
As discussed on the mailinglist, mdadm will be moving to creating
version 1.1 metadata by default, and we should be creating arrays with
bitmaps (except for /boot and swap).

So this patch adds parameters to mdcreate to specify wether to use a
bitmap or not and which metadata version to use.

It then changes MDRaidArrayDevice.create() to:
1) For /boot force metadata version to 1.0, as grub cannot handle 1.1,
   and don't use a bitmap.
2) For swap don't use a bitmap.
3) Everything else create with bitmap