Enable the .z part of the anaconda version number.
We are using an x.y version number scheme for anaconda in Fedora.  The x
value equals the Fedora version number (e.g., 13).  The y value is the
release of anaconda we've made.  We start at zero when we reset the x
value and increment y by one for each release.

Since RHEL anaconda branches are branched from Fedora, we use an x.y.z
version number.  In this case, the x and y values will never change and
reflect the release we branched from in Fedora.  The z value will start
at zero and increment with each release we make in this branch.

To ensure the first build we do is 13.21.0, set the z value to -1 so
make bumpver works.