Reset network setting input counters for IPv4 and IPv6 (#553761).
In loader, the manual TCP/IP configuration loop prompts the user then
validates the input.  Each time a particular piece of input validates,
the counter is incremented.  If the counter does not equal the right now
at the end of the loop, we tell the user to re-enter the information
correctly.  But the second time around, if any input was valid the first
time, it will be counted again, throwing off the count we check at the
end of the loop.

This patch just resets the IPv4 and IPv6 counters if we detected invalid
input.