Support ignore all/reinit all on the disk reinitialization question (#512011).
It's possible for the user to have a huge number of disks, which could result in
this dialog popping up a huge number of times.  Therefore, it'd be a good idea
to allow the user to apply their same choice to all following times when the
dialog would pop up.

Note that it would be nicer to add some sort of checkbox to MessageWindow to
make this more clear, but that is not exactly trivial to do.