Fix visibility counting on filter notebook pages.
Since all notebook pages share the same store, a device being visible on one
page means it'll be counted as being visible on all pages.  That means
flipping to the correct initial page doesn't work.  Instead, we need to count
using the visible methods on callback objects to only count the devices that
are members of each page as visible.