Do not reference Fedora bugs in rhel6-branch commits.
We are at the point where only RHEL-6 bugs can be referenced in commit
messages to rhel6-branch.  If we reference Fedora bugs, it screws with
the dist-cvs commit hooks.

If you are fixing a Fedora bug that also applies to RHEL-6, please clone
the Fedora bug for RHEL-6 and reference that on rhel6-branch.

I changed 558177 to 558689 and updated the commit message as well.