Add pure python EDD code parsing and compareDrives substitute (#478996)
There are various issues with the EDD C-code, mostly that it is way
too complicated for what it does. This patch re-implements it in
python, and adds a compareDrives substitute to the storage class.

Note that the compareDrives substitute is a copy and paste job of the
isys code (using the new storage edd_dict). The isys code will be removed
in a separate patch.