Remove /sys prefix in udev_enumerate_devices()
The recent base_udev.py changes to switch over to
global_udev.enumerate_devices(), move the removing of the /sys prefix to
udev_get_device(). But in various places in the devicetree we call
udev_get_device() (through udev_get_block_device) to get information from
parents which are not yet in the tree (ie when we scan a partition
before its disk).

All these calls are broken now, as all those pass in a path without /sys
at the front and the new udev_get_device() wants it with /sys in front.

This patch makes udev_enumerate_devices() return paths without the /sys
prefix, and reverses the changes to udev_get_device() to expect a full path,
so that we use sysfs paths without /sys in front consistently everywhere.