pyudev: explicitly specify all return value and argument types (#559394)
We were relying on the ctypes behavior to default to int as return type for
all functions. But most of the functions in question are not returning int's
but pointers, which are 64 bit on s390x (and ppc64 and x86_64).

Due to the fact that we were getting virtual addresses below 4G + the C
calling conventions on x86_64 making it not matter when passing in an 32 bit
int instead of a 64 bit one, as long as the 32 bit one represent the correct
number, we got away with this on x86_64 (unfortunately really).

This patch very likely also fixes #560035.

Note I also threw in a traceback fix in the same codepath, baseudev assumes
that devices always have a DEVLINKS property, which is not true for for
example network devices.