Fixup partition aligning (#560586)
Our partitioning code was assuming that having a partition from
sector 0 - 512 means the partition stops at sector 512, iow that
it goes from sector 0 till 511, but that is not how parted geometries
work, the range includes the end sector. So 0 - 512 will get you
a 513 sectors partition.

This patch adds an endAligment property to disklabel which can be used
to properly align end sectors (which must be aligned to the disklabel's
alignment with the offset decreased by 1), patches partitioning.py to
use this to align end sectors and fixes up end sector calculations in
partitioning.py to be start + length - 1.

On DASD's (s390), the wrong end sector was causing a traceback (#560586),
as DASD's disklabels need partitions to be cylinder aligned, and the wrong
end sector violated this.